Extract methods to tab helper
