+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2016-11-08 12:47:19 +0300 using RuboCop version 0.45.0.
+# on 2016-11-08 13:05:06 +0300 using RuboCop version 0.45.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -23,7 +23,7 @@ Metrics/AbcSize:
 # Offense count: 1
 # Configuration parameters: CountComments.
 Metrics/ClassLength:
-  Max: 704
+  Max: 641
 
 # Offense count: 3
 Metrics/CyclomaticComplexity:
@@ -59,6 +59,7 @@ Style/AccessorMethodName:
   Exclude:
     - 'lib/onlyoffice_webdriver_wrapper/amazon_s3_wrapper.rb'
     - 'lib/onlyoffice_webdriver_wrapper/webdriver.rb'
+    - 'lib/onlyoffice_webdriver_wrapper/webdriver/webdriver_tab_helper.rb'
 
 # Offense count: 4
 Style/Documentation:
